requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
cycler==0.10.0
Cython==0.24.1
decorator==4.0.10
Django==1.10.1
django-filter==1.0.1
django-model-utils==2.6
djangorestframework==3.5.3
enum34==1.1.6
factory-boy==1.3.0
funcsigs==1.0.2
httpie==0.9.9
ipython==5.1.0
ipython-genutils==0.1.0
lxml==3.6.4
Markdown==2.6.8
matplotlib==1.5.3
mock==2.0.0
numpy==1.11.1
pandas==0.19.0rc1
pathlib2==2.1.0
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
pocketsphinx==0.1.3
prompt-toolkit==1.0.7
protobuf==3.0.0
ptyprocess==0.5.1
PyAudio==0.2.10
Pygments==2.1.3
pyparsing==2.1.9
python-dateutil==2.5.3
pytz==2016.6.1
requests==2.11.1
scikit-learn==0.17.1
scipy==0.18.0
simplegeneric==0.8.1
six==1.10.0
SpeechRecognition==3.6.3
tensorflow==0.11.0rc0
termcolor==1.1.0
Theano==0.8.2
traitlets==4.3.0
vboxapi==1.0
virtualenv==15.1.0
wcwidth==0.1.7
word2vec==0.9.1
-e git+https://github.com/dmlc/xgboost@bb388cbb31ddc0284ab0469bfdd6913aed5ad23b#egg=xgboost&subdirectory=python-package
